split profile
