+++ shared/actions/json/profile.json
@@ -1,6 +1,6 @@
 {
   "prelude": [
-    "import * as Constants from '../constants/profile'"
+    "import * as Types from '../constants/types/profile'"
   ],
   "actions": {
     "addProof": {
@@ -56,9 +56,9 @@
       "errorCode?": "?number"
     },
     "updatePgpInfo": {
-      "info": "$Shape<Constants.PgpInfo>",
+      "info": "$Shape<Types.PgpInfo>",
       "canError": {
-        "error": "Constants.PgpInfoError"
+        "error": "Types.PgpInfoError"
       }
     },
     "updatePgpPublicKey": {
