+++ shared/actions/profile/index.js
@@ -1,5 +1,6 @@
 // @flow
 import * as AppGen from '../app-gen'
+import * as Types from '../../constants/types/profile'
 import * as Constants from '../../constants/profile'
 import * as ProfileGen from '../profile-gen'
 import * as Saga from '../../util/saga'
@@ -167,7 +168,7 @@ function* _onAppLink(action: AppGen.LinkPayload): Saga.SagaGenerator<any, any> {
 
 function* _outputInstructionsActionLink(): Saga.SagaGenerator<any, any> {
   const getProfile = (state: TypedState) => state.profile
-  const profile: Constants.State = (yield Saga.select(getProfile): any)
+  const profile: Types.State = (yield Saga.select(getProfile): any)
   switch (profile.platform) {
     case 'twitter':
       yield Saga.call(
