+++ shared/constants/reducer.js
@@ -12,7 +12,7 @@ import type {State as Login} from '../constants/types/login'
 import type {State as Notification} from '../constants/types/notifications'
 import type {State as Pgp} from '../constants/types/pgp'
 import type {State as Pinentry} from '../constants/types/pinentry'
-import type {State as Profile} from '../constants/profile'
+import type {State as Profile} from '../constants/types/profile'
 import type {State as Push} from '../constants/push'
 import type {State as RouteTree} from '../constants/route-tree'
 import type {State as Settings} from '../constants/settings'
