+++ shared/reducers/profile.js
@@ -1,8 +1,9 @@
 // @flow
 import * as ProfileGen from '../actions/profile-gen'
+import * as Types from '../constants/types/profile'
 import * as Constants from '../constants/profile'
 
-export default function(state: Constants.State = Constants.initialState, action: ProfileGen.Actions) {
+export default function(state: Types.State = Constants.initialState, action: ProfileGen.Actions) {
   switch (action.type) {
     case ProfileGen.resetStore:
       return {...Constants.initialState}
